+++ git_machete/tests/functional/test_machete.py
@@ -12,15 +12,9 @@ import pytest  # type: ignore
 from git_machete import cli
 from git_machete.exceptions import MacheteException
 from git_machete.git_operations import LocalBranchShortName
-from git_machete.github import get_parsed_github_remote_url
-from git_machete.tests.functional.commons import (FAKE_GITHUB_REMOTE_PATTERNS,
-                                                  FakeCommandLineOptions,
-                                                  GitRepositorySandbox,
-                                                  MockContextManager,
-                                                  MockGitHubAPIState,
-                                                  MockHTTPError,
+from git_machete.tests.functional.commons import (GitRepositorySandbox,
                                                   get_current_commit_hash, git,
-                                                  mock_run_cmd, popen)
+                                                  mock_run_cmd)
 from git_machete.utils import dim
 
 
@@ -1102,790 +1096,3 @@ class TestMachete:
 
         with pytest.raises(SystemExit):
             self.launch_command("advance", '-y')
-
-    git_api_state_for_test_retarget_pr = MockGitHubAPIState(
-        [{'head': {'ref': 'feature', 'repo': mock_repository_info}, 'user': {'login': 'github_user'}, 'base': {'ref': 'root'}, 'number': '15',
-          'html_url': 'www.github.com', 'state': 'open'}])
-
-    @mock.patch('git_machete.utils.run_cmd', mock_run_cmd)  # to hide git outputs in tests
-    @mock.patch('urllib.request.Request', git_api_state_for_test_retarget_pr.new_request())
-    @mock.patch('urllib.request.urlopen', MockContextManager)
-    def test_retarget_pr(self) -> None:
-        branchs_first_commit_msg = "First commit on branch."
-        branchs_second_commit_msg = "Second commit on branch."
-        (
-            self.repo_sandbox.new_branch("root")
-                .commit("First commit on root.")
-                .new_branch("branch-1")
-                .commit(branchs_first_commit_msg)
-                .commit(branchs_second_commit_msg)
-                .push()
-                .new_branch('feature')
-                .commit('introduce feature')
-                .push()
-                .check_out('feature')
-                .add_remote('new_origin', 'https://github.com/user/repo.git')
-        )
-
-        self.launch_command("discover", "-y")
-        self.assert_command(['github', 'retarget-pr'], 'The base branch of PR #15 has been switched to `branch-1`\n', strip_indentation=False)
-        self.assert_command(['github', 'retarget-pr'], 'The base branch of PR #15 is already `branch-1`\n', strip_indentation=False)
-
-    git_api_state_for_test_anno_prs = MockGitHubAPIState([
-        {'head': {'ref': 'ignore-trailing', 'repo': mock_repository_info}, 'user': {'login': 'github_user'}, 'base': {'ref': 'hotfix/add-trigger'}, 'number': '3', 'html_url': 'www.github.com', 'state': 'open'},
-        {'head': {'ref': 'allow-ownership-link', 'repo': mock_repository_info}, 'user': {'login': 'github_user'}, 'base': {'ref': 'develop'}, 'number': '7', 'html_url': 'www.github.com', 'state': 'open'},
-        {'head': {'ref': 'call-ws', 'repo': mock_repository_info}, 'user': {'login': 'github_user'}, 'base': {'ref': 'develop'}, 'number': '31', 'html_url': 'www.github.com', 'state': 'open'}
-    ])
-
-    @mock.patch('git_machete.utils.run_cmd', mock_run_cmd)  # to hide git outputs in tests
-    @mock.patch('git_machete.github.derive_current_user_login', mock_derive_current_user_login)
-    @mock.patch('urllib.request.urlopen', MockContextManager)
-    @mock.patch('urllib.request.Request', git_api_state_for_test_anno_prs.new_request())
-    def test_anno_prs(self) -> None:
-        (
-            self.repo_sandbox.new_branch("root")
-                .commit("root")
-                .new_branch("develop")
-                .commit("develop commit")
-                .new_branch("allow-ownership-link")
-                .commit("Allow ownership links")
-                .push()
-                .new_branch("build-chain")
-                .commit("Build arbitrarily long chains")
-                .check_out("allow-ownership-link")
-                .commit("1st round of fixes")
-                .check_out("develop")
-                .commit("Other develop commit")
-                .push()
-                .new_branch("call-ws")
-                .commit("Call web service")
-                .commit("1st round of fixes")
-                .push()
-                .new_branch("drop-constraint")
-                .commit("Drop unneeded SQL constraints")
-                .check_out("call-ws")
-                .commit("2nd round of fixes")
-                .check_out("root")
-                .new_branch("master")
-                .commit("Master commit")
-                .push()
-                .new_branch("hotfix/add-trigger")
-                .commit("HOTFIX Add the trigger")
-                .push()
-                .commit_amend("HOTFIX Add the trigger (amended)")
-                .new_branch("ignore-trailing")
-                .commit("Ignore trailing data")
-                .sleep(1)
-                .commit_amend("Ignore trailing data (amended)")
-                .push()
-                .reset_to("ignore-trailing@{1}")
-                .delete_branch("root")
-                .add_remote('new_origin', 'https://github.com/user/repo.git')
-        )
-        self.launch_command("discover", "-y")
-        self.launch_command('github', 'anno-prs')
-        self.assert_command(
-            ["status"],
-            """
-            master
-            |
-            o-hotfix/add-trigger (diverged from origin)
-              |
-              o-ignore-trailing *  PR #3 (github_user) (diverged from & older than origin)
-
-            develop
-            |
-            x-allow-ownership-link  PR #7 (github_user) (ahead of origin)
-            | |
-            | x-build-chain (untracked)
-            |
-            o-call-ws  PR #31 (github_user) (ahead of origin)
-              |
-              x-drop-constraint (untracked)
-            """,
-        )
-
-    git_api_state_for_test_create_pr = MockGitHubAPIState([{'head': {'ref': 'ignore-trailing', 'repo': mock_repository_info}, 'user': {'login': 'github_user'}, 'base': {'ref': 'hotfix/add-trigger'}, 'number': '3', 'html_url': 'www.github.com', 'state': 'open'}],
-                                                          issues=[{'number': '4'}, {'number': '5'}, {'number': '6'}])
-
-    @mock.patch('git_machete.cli.exit_script', mock_exit_script)
-    @mock.patch('git_machete.client.MacheteClient.ask_if', mock_ask_if)
-    # We need to mock GITHUB_REMOTE_PATTERNS in the tests for `test_github_create_pr` due to `git fetch` executed by `create-pr` subcommand.
-    @mock.patch('git_machete.github.GITHUB_REMOTE_PATTERNS', FAKE_GITHUB_REMOTE_PATTERNS)
-    @mock.patch('git_machete.utils.run_cmd', mock_run_cmd)  # to hide git outputs in tests
-    @mock.patch('git_machete.options.CommandLineOptions', FakeCommandLineOptions)
-    @mock.patch('urllib.error.HTTPError', MockHTTPError)  # need to provide read() method, which does not actually reads error from url
-    @mock.patch('urllib.request.Request', git_api_state_for_test_create_pr.new_request())
-    @mock.patch('urllib.request.urlopen', MockContextManager)
-    def test_github_create_pr(self) -> None:
-        (
-            self.repo_sandbox.new_branch("root")
-                .commit("initial commit")
-                .new_branch("develop")
-                .commit("first commit")
-                .new_branch("allow-ownership-link")
-                .commit("Enable ownership links")
-                .push()
-                .new_branch("build-chain")
-                .commit("Build arbitrarily long chains of PRs")
-                .check_out("allow-ownership-link")
-                .commit("fixes")
-                .check_out("develop")
-                .commit("Other develop commit")
-                .push()
-                .new_branch("call-ws")
-                .commit("Call web service")
-                .commit("1st round of fixes")
-                .push()
-                .new_branch("drop-constraint")
-                .commit("Drop unneeded SQL constraints")
-                .check_out("call-ws")
-                .commit("2nd round of fixes")
-                .check_out("root")
-                .new_branch("master")
-                .commit("Master commit")
-                .push()
-                .new_branch("hotfix/add-trigger")
-                .commit("HOTFIX Add the trigger")
-                .push()
-                .commit_amend("HOTFIX Add the trigger (amended)")
-                .new_branch("ignore-trailing")
-                .commit("Ignore trailing data")
-                .sleep(1)
-                .commit_amend("Ignore trailing data (amended)")
-                .push()
-                .reset_to("ignore-trailing@{1}")
-                .delete_branch("root")
-                .new_branch('chore/fields')
-                .commit("remove outdated fields")
-                .check_out("call-ws")
-                .add_remote('new_origin', 'https://github.com/user/repo.git')
-        )
-
-        self.launch_command("discover")
-        self.launch_command("github", "create-pr")
-        # ahead of origin state, push is advised and accepted
-        self.assert_command(
-            ['status'],
-            """
-            master
-            |
-            o-hotfix/add-trigger (diverged from origin)
-              |
-              o-ignore-trailing (diverged from & older than origin)
-                |
-                o-chore/fields (untracked)
-
-            develop
-            |
-            x-allow-ownership-link (ahead of origin)
-            | |
-            | x-build-chain (untracked)
-            |
-            o-call-ws *  PR #4
-              |
-              x-drop-constraint (untracked)
-            """,
-        )
-        self.repo_sandbox.check_out('chore/fields')
-        #  untracked state (can only create pr when branch is pushed)
-        self.launch_command("github", "create-pr", "--draft")
-        self.assert_command(
-            ['status'],
-            """
-            master
-            |
-            o-hotfix/add-trigger (diverged from origin)
-              |
-              o-ignore-trailing (diverged from & older than origin)
-                |
-                o-chore/fields *  PR #5
-
-            develop
-            |
-            x-allow-ownership-link (ahead of origin)
-            | |
-            | x-build-chain (untracked)
-            |
-            o-call-ws  PR #4
-              |
-              x-drop-constraint (untracked)
-            """,
-        )
-
-        (
-            self.repo_sandbox.check_out('hotfix/add-trigger')
-            .commit('trigger released')
-            .commit('minor changes applied')
-        )
-
-        # diverged from and newer than origin
-        self.launch_command("github", "create-pr")
-        self.assert_command(
-            ['status'],
-            """
-            master
-            |
-            o-hotfix/add-trigger *  PR #6
-              |
-              x-ignore-trailing (diverged from & older than origin)
-                |
-                o-chore/fields  PR #5
-
-            develop
-            |
-            x-allow-ownership-link (ahead of origin)
-            | |
-            | x-build-chain (untracked)
-            |
-            o-call-ws  PR #4
-              |
-              x-drop-constraint (untracked)
-            """,
-        )
-        expected_error_message = "A pull request already exists for test_repo:hotfix/add-trigger."
-        with pytest.raises(MacheteException) as e:
-            self.launch_command("github", "create-pr")
-        if e:
-            assert e.value.args[0] == expected_error_message, 'Verify that expected error message has appeared when given pull request to create is already created.'
-
-        # check against head branch is ancestor or equal to base branch
-        (
-            self.repo_sandbox.check_out('develop')
-            .new_branch('testing/endpoints')
-            .push()
-        )
-        self.launch_command('discover')
-
-        expected_error_message = "All commits in `testing/endpoints` branch are already included in `develop` branch.\nCannot create pull request."
-        with pytest.raises(MacheteException) as e:
-            self.launch_command("github", "create-pr")
-        if e:
-            assert e.value.parameter == expected_error_message, 'Verify that expected error message has appeared when head branch is equal or ancestor of base branch.'
-
-        self.repo_sandbox.check_out('develop')
-        expected_error_message = "Branch `develop` does not have a parent branch (it is a root), base branch for the PR cannot be established."
-        with pytest.raises(MacheteException) as e:
-            self.launch_command("github", "create-pr")
-        if e:
-            assert e.value.parameter == expected_error_message, 'Verify that expected error message has appeared when creating PR from root branch.'
-
-    git_api_state_for_test_create_pr_missing_base_branch_on_remote = MockGitHubAPIState([{'head': {'ref': 'chore/redundant_checks', 'repo': mock_repository_info}, 'user': {'login': 'github_user'}, 'base': {'ref': 'restrict_access'}, 'number': '18', 'html_url': 'www.github.com', 'state': 'open'}])
-
-    # We need to mock GITHUB_REMOTE_PATTERNS in the tests for `test_github_create_pr` due to `git fetch` executed by `create-pr` subcommand.
-    @mock.patch('git_machete.github.GITHUB_REMOTE_PATTERNS', FAKE_GITHUB_REMOTE_PATTERNS)
-    @mock.patch('git_machete.github.__get_github_token', mock__get_github_token)
-    @mock.patch('git_machete.utils.run_cmd', mock_run_cmd)  # to hide git outputs in tests
-    @mock.patch('git_machete.options.CommandLineOptions', FakeCommandLineOptions)
-    @mock.patch('git_machete.client.MacheteClient.ask_if', mock_ask_if)
-    @mock.patch('urllib.request.urlopen', MockContextManager)
-    @mock.patch('urllib.request.Request', git_api_state_for_test_create_pr_missing_base_branch_on_remote.new_request())
-    def test_github_create_pr_missing_base_branch_on_remote(self) -> None:
-        (
-            self.repo_sandbox.new_branch("root")
-                .commit("initial commit")
-                .new_branch("develop")
-                .commit("first commit on develop")
-                .push()
-                .new_branch("feature/api_handling")
-                .commit("Introduce GET and POST methods on API")
-                .new_branch("feature/api_exception_handling")
-                .commit("catch exceptions coming from API")
-                .push()
-                .delete_branch("root")
-        )
-
-        self.launch_command('discover')
-
-        expected_msg = ("Fetching origin...\n"
-                        "Warn: Base branch for this PR (`feature/api_handling`) is not found on remote, pushing...\n"
-                        "Creating a PR from `feature/api_exception_handling` to `feature/api_handling`... OK, see www.github.com\n")
-        self.assert_command(['github', 'create-pr'], expected_msg, strip_indentation=False)
-        self.assert_command(
-            ['status'],
-            """
-            develop
-            |
-            o-feature/api_handling
-              |
-              o-feature/api_exception_handling *  PR #19
-            """,
-        )
-
-    git_api_state_for_test_checkout_prs = MockGitHubAPIState([
-        {'head': {'ref': 'chore/redundant_checks', 'repo': mock_repository_info}, 'user': {'login': 'github_user'}, 'base': {'ref': 'restrict_access'}, 'number': '18', 'html_url': 'www.github.com', 'state': 'open'},
-        {'head': {'ref': 'restrict_access', 'repo': mock_repository_info}, 'user': {'login': 'github_user'}, 'base': {'ref': 'allow-ownership-link'}, 'number': '17', 'html_url': 'www.github.com', 'state': 'open'},
-        {'head': {'ref': 'allow-ownership-link', 'repo': mock_repository_info}, 'user': {'login': 'github_user'}, 'base': {'ref': 'bugfix/feature'}, 'number': '12', 'html_url': 'www.github.com', 'state': 'open'},
-        {'head': {'ref': 'bugfix/feature', 'repo': mock_repository_info}, 'user': {'login': 'github_user'}, 'base': {'ref': 'enhance/feature'}, 'number': '6', 'html_url': 'www.github.com', 'state': 'open'},
-        {'head': {'ref': 'enhance/add_user', 'repo': mock_repository_info}, 'user': {'login': 'github_user'}, 'base': {'ref': 'develop'}, 'number': '19', 'html_url': 'www.github.com', 'state': 'open'},
-        {'head': {'ref': 'testing/add_user', 'repo': mock_repository_info}, 'user': {'login': 'github_user'}, 'base': {'ref': 'bugfix/add_user'}, 'number': '22', 'html_url': 'www.github.com', 'state': 'open'},
-        {'head': {'ref': 'chore/comments', 'repo': mock_repository_info}, 'user': {'login': 'github_user'}, 'base': {'ref': 'testing/add_user'}, 'number': '24', 'html_url': 'www.github.com', 'state': 'open'},
-        {'head': {'ref': 'ignore-trailing', 'repo': mock_repository_info}, 'user': {'login': 'github_user'}, 'base': {'ref': 'hotfix/add-trigger'}, 'number': '3', 'html_url': 'www.github.com', 'state': 'open'},
-        {'head': {'ref': 'bugfix/remove-n-option', 'repo': {'full_name': 'testing/checkout_prs', 'html_url': GitRepositorySandbox.second_remote_path}}, 'user': {'login': 'github_user'}, 'base': {'ref': 'develop'}, 'number': '5', 'html_url': 'www.github.com', 'state': 'closed'}
-    ])
-
-    @mock.patch('git_machete.cli.exit_script', mock_exit_script)
-    # We need to mock GITHUB_REMOTE_PATTERNS in the tests for `test_github_checkout_prs` due to `git fetch` executed by `checkout-prs` subcommand.
-    @mock.patch('git_machete.github.GITHUB_REMOTE_PATTERNS', FAKE_GITHUB_REMOTE_PATTERNS)
-    @mock.patch('git_machete.options.CommandLineOptions', FakeCommandLineOptions)
-    @mock.patch('git_machete.utils.run_cmd', mock_run_cmd)  # to hide git outputs in tests
-    @mock.patch('git_machete.github.__get_github_token', mock__get_github_token)
-    @mock.patch('urllib.request.Request', git_api_state_for_test_checkout_prs.new_request())
-    @mock.patch('urllib.request.urlopen', MockContextManager)
-    def test_github_checkout_prs(self) -> None:
-        (
-            self.repo_sandbox.new_branch("root")
-            .commit("initial commit")
-            .new_branch("develop")
-            .commit("first commit")
-            .push()
-            .new_branch("enhance/feature")
-            .commit("introduce feature")
-            .push()
-            .new_branch("bugfix/feature")
-            .commit("bugs removed")
-            .push()
-            .new_branch("allow-ownership-link")
-            .commit("fixes")
-            .push()
-            .new_branch('restrict_access')
-            .commit('authorized users only')
-            .push()
-            .new_branch("chore/redundant_checks")
-            .commit('remove some checks')
-            .push()
-            .check_out("root")
-            .new_branch("master")
-            .commit("Master commit")
-            .push()
-            .new_branch("hotfix/add-trigger")
-            .commit("HOTFIX Add the trigger")
-            .push()
-            .new_branch("ignore-trailing")
-            .commit("Ignore trailing data")
-            .push()
-            .delete_branch("root")
-            .new_branch('chore/fields')
-            .commit("remove outdated fields")
-            .push()
-            .check_out('develop')
-            .new_branch('enhance/add_user')
-            .commit('allow externals to add users')
-            .push()
-            .new_branch('bugfix/add_user')
-            .commit('first round of fixes')
-            .push()
-            .new_branch('testing/add_user')
-            .commit('add test set for add_user feature')
-            .push()
-            .new_branch('chore/comments')
-            .commit('code maintenance')
-            .push()
-            .check_out('master')
-        )
-        for branch in ('chore/redundant_checks', 'restrict_access', 'allow-ownership-link', 'bugfix/feature', 'enhance/add_user', 'testing/add_user', 'chore/comments', 'bugfix/add_user'):
-            self.repo_sandbox.execute(f"git branch -D {branch}")
-
-        self.launch_command('discover')
-
-        # not broken chain of pull requests (root found in dependency tree)
-        self.launch_command('github', 'checkout-prs', '18')
-        self.assert_command(
-            ["status"],
-            """
-            master
-            |
-            o-hotfix/add-trigger
-              |
-              o-ignore-trailing  PR #3 (github_user)
-                |
-                o-chore/fields
-
-            develop
-            |
-            o-enhance/feature
-              |
-              o-bugfix/feature  PR #6 (github_user)
-                |
-                o-allow-ownership-link  PR #12 (github_user)
-                  |
-                  o-restrict_access  PR #17 (github_user)
-                    |
-                    o-chore/redundant_checks *  PR #18 (github_user)
-            """
-        )
-        # broken chain of pull requests (add new root)
-        self.launch_command('github', 'checkout-prs', '24')
-        self.assert_command(
-            ["status"],
-            """
-            master
-            |
-            o-hotfix/add-trigger
-              |
-              o-ignore-trailing  PR #3 (github_user)
-                |
-                o-chore/fields
-
-            develop
-            |
-            o-enhance/feature
-              |
-              o-bugfix/feature  PR #6 (github_user)
-                |
-                o-allow-ownership-link  PR #12 (github_user)
-                  |
-                  o-restrict_access  PR #17 (github_user)
-                    |
-                    o-chore/redundant_checks  PR #18 (github_user)
-
-            bugfix/add_user
-            |
-            o-testing/add_user  PR #22 (github_user)
-              |
-              o-chore/comments *  PR #24 (github_user)
-            """
-        )
-
-        # broken chain of pull requests (branches already added)
-        self.launch_command('github', 'checkout-prs', '24')
-        self.assert_command(
-            ["status"],
-            """
-            master
-            |
-            o-hotfix/add-trigger
-              |
-              o-ignore-trailing  PR #3 (github_user)
-                |
-                o-chore/fields
-
-            develop
-            |
-            o-enhance/feature
-              |
-              o-bugfix/feature  PR #6 (github_user)
-                |
-                o-allow-ownership-link  PR #12 (github_user)
-                  |
-                  o-restrict_access  PR #17 (github_user)
-                    |
-                    o-chore/redundant_checks  PR #18 (github_user)
-
-            bugfix/add_user
-            |
-            o-testing/add_user  PR #22 (github_user)
-              |
-              o-chore/comments *  PR #24 (github_user)
-            """
-        )
-
-        # all PRs
-        self.launch_command('github', 'checkout-prs', '--all')
-        self.assert_command(
-            ["status"],
-            """
-            master
-            |
-            o-hotfix/add-trigger
-              |
-              o-ignore-trailing  PR #3 (github_user)
-                |
-                o-chore/fields
-
-            develop
-            |
-            o-enhance/feature
-            | |
-            | o-bugfix/feature  PR #6 (github_user)
-            |   |
-            |   o-allow-ownership-link  PR #12 (github_user)
-            |     |
-            |     o-restrict_access  PR #17 (github_user)
-            |       |
-            |       o-chore/redundant_checks  PR #18 (github_user)
-            |
-            o-enhance/add_user  PR #19 (github_user)
-
-            bugfix/add_user
-            |
-            o-testing/add_user  PR #22 (github_user)
-              |
-              o-chore/comments *  PR #24 (github_user)
-            """
-        )
-
-        # check against wrong pr number
-        repo: str
-        org: str
-        (org, repo) = get_parsed_github_remote_url(self.repo_sandbox.remote_path)
-        expected_error_message = f"PR #100 is not found in repository `{org}/{repo}`"
-        with pytest.raises(MacheteException) as e:
-            self.launch_command('github', 'checkout-prs', '100')
-        if e:
-            assert e.value.parameter == expected_error_message, \
-                'Verify that expected error message has appeared when given pull request to checkout does not exists.'
-
-        with pytest.raises(MacheteException) as e:
-            self.launch_command('github', 'checkout-prs', '19', '100')
-        if e:
-            assert e.value.parameter == expected_error_message, \
-                'Verify that expected error message has appeared when one of the given pull requests to checkout does not exists.'
-
-        # check against user with no open pull requests
-        expected_msg = ("Checking for open GitHub PRs...\n"
-                        f"Warn: User `tester` has no open pull request in repository `{org}/{repo}`\n")
-        self.assert_command(['github', 'checkout-prs', '--by', 'tester'], expected_msg, strip_indentation=False)
-
-        # Check against closed pull request with head branch deleted from remote
-        local_path = popen("mktemp -d")
-        self.repo_sandbox.new_repo(GitRepositorySandbox.second_remote_path)
-        (self.repo_sandbox.new_repo(local_path)
-            .execute(f"git remote add origin {GitRepositorySandbox.second_remote_path}")
-            .execute('git config user.email "tester@test.com"')
-            .execute('git config user.name "Tester Test"')
-            .new_branch('main')
-            .commit('initial commit')
-            .push()
-         )
-        os.chdir(self.repo_sandbox.local_path)
-
-        expected_error_message = "Could not check out PR #5 because its head branch `bugfix/remove-n-option` is already deleted from `testing`."
-        with pytest.raises(MacheteException) as e:
-            self.launch_command('github', 'checkout-prs', '5')
-        if e:
-            assert e.value.parameter == expected_error_message, \
-                'Verify that expected error message has appeared when given pull request to checkout have already deleted branch from remote.'
-
-        # Check against pr come from fork
-        os.chdir(local_path)
-        (self.repo_sandbox
-         .new_branch('bugfix/remove-n-option')
-         .commit('first commit')
-         .push()
-         )
-        os.chdir(self.repo_sandbox.local_path)
-
-        expected_msg = ("Checking for open GitHub PRs...\n"
-                        "Warn: Pull request #5 is already closed.\n"
-                        "Pull request `#5` checked out at local branch `bugfix/remove-n-option`\n")
-
-        self.assert_command(['github', 'checkout-prs', '5'], expected_msg, strip_indentation=False)
-
-        # Check against multiple PRs
-        expected_msg = 'Checking for open GitHub PRs...\n'
-
-        self.assert_command(['github', 'checkout-prs', '3', '12'], expected_msg, strip_indentation=False)
-
-    git_api_state_for_test_github_checkout_prs_fresh_repo = MockGitHubAPIState([
-        {'head': {'ref': 'comments/add_docstrings', 'repo': mock_repository_info}, 'user': {'login': 'github_user'}, 'base': {'ref': 'improve/refactor'}, 'number': '2', 'html_url': 'www.github.com', 'state': 'open'},
-        {'head': {'ref': 'restrict_access', 'repo': mock_repository_info}, 'user': {'login': 'github_user'}, 'base': {'ref': 'allow-ownership-link'}, 'number': '17', 'html_url': 'www.github.com', 'state': 'open'},
-        {'head': {'ref': 'improve/refactor', 'repo': mock_repository_info}, 'user': {'login': 'github_user'}, 'base': {'ref': 'chore/sync_to_docs'}, 'number': '1', 'html_url': 'www.github.com', 'state': 'open'},
-        {'head': {'ref': 'sphinx_export', 'repo': {'full_name': 'testing/checkout_prs', 'html_url': GitRepositorySandbox.second_remote_path}}, 'user': {'login': 'github_user'}, 'base': {'ref': 'comments/add_docstrings'}, 'number': '23', 'html_url': 'www.github.com', 'state': 'closed'}
-    ])
-
-    @mock.patch('git_machete.utils.run_cmd', mock_run_cmd)  # to hide git outputs in tests
-    # We need to mock GITHUB_REMOTE_PATTERNS in the tests for `test_github_checkout_prs_freshly_cloned` due to `git fetch` executed by `checkout-prs` subcommand.
-    @mock.patch('git_machete.options.CommandLineOptions', FakeCommandLineOptions)
-    @mock.patch('git_machete.github.GITHUB_REMOTE_PATTERNS', FAKE_GITHUB_REMOTE_PATTERNS)
-    @mock.patch('urllib.request.urlopen', MockContextManager)
-    @mock.patch('urllib.request.Request', git_api_state_for_test_github_checkout_prs_fresh_repo.new_request())
-    def test_github_checkout_prs_freshly_cloned(self) -> None:
-        (
-            self.repo_sandbox.new_branch("root")
-            .commit("initial commit")
-            .new_branch("develop")
-            .commit("first commit")
-            .push()
-            .new_branch("chore/sync_to_docs")
-            .commit("synchronize docs")
-            .push()
-            .new_branch("improve/refactor")
-            .commit("refactor code")
-            .push()
-            .new_branch("comments/add_docstrings")
-            .commit("docstring added")
-            .push()
-            .new_branch("sphinx_export")
-            .commit("export docs to html")
-            .push()
-            .check_out("root")
-            .new_branch("master")
-            .commit("Master commit")
-            .push()
-            .delete_branch("root")
-            .push()
-        )
-        for branch in ('develop', 'chore/sync_to_docs', 'improve/refactor', 'comments/add_docstrings'):
-            self.repo_sandbox.execute(f"git branch -D {branch}")
-        local_path = popen("mktemp -d")
-        os.chdir(local_path)
-        self.repo_sandbox.execute(f'git clone {self.repo_sandbox.remote_path}')
-        os.chdir(os.path.join(local_path, os.listdir()[0]))
-
-        for branch in ('develop', 'chore/sync_to_docs', 'improve/refactor', 'comments/add_docstrings'):
-            self.repo_sandbox.execute(f"git branch -D -r origin/{branch}")
-
-        local_path = popen("mktemp -d")
-        self.repo_sandbox.new_repo(GitRepositorySandbox.second_remote_path)
-        (
-            self.repo_sandbox.new_repo(local_path)
-            .execute(f"git remote add origin {GitRepositorySandbox.second_remote_path}")
-            .execute('git config user.email "tester@test.com"')
-            .execute('git config user.name "Tester Test"')
-            .new_branch('feature')
-            .commit('initial commit')
-            .push()
-        )
-        os.chdir(self.repo_sandbox.local_path)
-        self.rewrite_definition_file("master")
-        expected_msg = ("Checking for open GitHub PRs...\n"
-                        "Pull request `#2` checked out at local branch `comments/add_docstrings`\n")
-        self.assert_command(
-            ['github', 'checkout-prs', '2'],
-            expected_msg,
-            strip_indentation=False
-        )
-
-        self.assert_command(
-            ["status"],
-            """
-            master
-
-            chore/sync_to_docs
-            |
-            o-improve/refactor  PR #1 (github_user)
-              |
-              o-comments/add_docstrings *  PR #2 (github_user)
-            """
-        )
-
-        # Check against closed pull request
-        self.repo_sandbox.execute('git branch -D sphinx_export')
-        expected_msg = ("Checking for open GitHub PRs...\n"
-                        "Warn: Pull request #23 is already closed.\n"
-                        "Pull request `#23` checked out at local branch `sphinx_export`\n")
-
-        self.assert_command(
-            ['github', 'checkout-prs', '23'],
-            expected_msg,
-            strip_indentation=False
-        )
-        self.assert_command(
-            ["status"],
-            """
-            master
-
-            chore/sync_to_docs
-            |
-            o-improve/refactor  PR #1 (github_user)
-              |
-              o-comments/add_docstrings  PR #2 (github_user)
-                |
-                o-sphinx_export *
-            """
-        )
-
-    git_api_state_for_test_github_checkout_prs_from_fork_with_deleted_repo = MockGitHubAPIState([
-        {'head': {'ref': 'feature/allow_checkout', 'repo': None}, 'user': {'login': 'github_user'}, 'base': {'ref': 'develop'}, 'number': '2', 'html_url': 'www.github.com', 'state': 'closed'},
-        {'head': {'ref': 'bugfix/allow_checkout', 'repo': mock_repository_info}, 'user': {'login': 'github_user'},
-         'base': {'ref': 'develop'}, 'number': '3', 'html_url': 'www.github.com', 'state': 'open'}
-    ])
-
-    @mock.patch('git_machete.git_operations.GitContext.fetch_ref', mock_fetch_ref)  # need to mock fetch_ref due to underlying `git fetch pull/head` calls
-    @mock.patch('git_machete.utils.run_cmd', mock_run_cmd)  # to hide git outputs in tests
-    # We need to mock GITHUB_REMOTE_PATTERNS in the tests for `test_github_checkout_prs_from_fork_with_deleted_repo` due to `git fetch` executed by `checkout-prs` subcommand.
-    @mock.patch('git_machete.options.CommandLineOptions', FakeCommandLineOptions)
-    @mock.patch('git_machete.github.GITHUB_REMOTE_PATTERNS', FAKE_GITHUB_REMOTE_PATTERNS)
-    @mock.patch('urllib.request.urlopen', MockContextManager)
-    @mock.patch('urllib.request.Request', git_api_state_for_test_github_checkout_prs_from_fork_with_deleted_repo.new_request())
-    def test_github_checkout_prs_from_fork_with_deleted_repo(self) -> None:
-        (
-            self.repo_sandbox.new_branch("root")
-            .commit('initial master commit')
-            .push()
-            .new_branch('develop')
-            .commit('initial develop commit')
-            .push()
-        )
-        self.launch_command('discover')
-        expected_msg = ("Checking for open GitHub PRs...\n"
-                        "Warn: Pull request #2 comes from fork and its repository is already deleted. No remote tracking data will be set up for `feature/allow_checkout` branch.\n"
-                        "Warn: Pull request #2 is already closed.\n"
-                        "Pull request `#2` checked out at local branch `feature/allow_checkout`\n")
-        self.assert_command(
-            ['github', 'checkout-prs', '2'],
-            expected_msg,
-            strip_indentation=False
-        )
-
-        assert 'feature/allow_checkout' == \
-            self.launch_command("show", "current").strip(), \
-            "Verify that 'git machete github checkout prs' performs 'git checkout' to " \
-            "the head branch of given pull request."
-
-    git_api_state_for_test_github_sync = MockGitHubAPIState([
-        {'head': {'ref': 'snickers', 'repo': mock_repository_info}, 'user': {'login': 'other_user'},
-         'base': {'ref': 'master'}, 'number': '7', 'html_url': 'www.github.com', 'state': 'open'}
-    ])
-
-    @mock.patch('git_machete.client.MacheteClient.should_perform_interactive_slide_out', mock_should_perform_interactive_slide_out)
-    @mock.patch('git_machete.utils.run_cmd', mock_run_cmd)
-    @mock.patch('git_machete.client.MacheteClient.ask_if', mock_ask_if)
-    @mock.patch('git_machete.options.CommandLineOptions', FakeCommandLineOptions)
-    @mock.patch('git_machete.github.GITHUB_REMOTE_PATTERNS', FAKE_GITHUB_REMOTE_PATTERNS)
-    @mock.patch('git_machete.github.__get_github_token', mock__get_github_token_fake)
-    @mock.patch('urllib.request.urlopen', MockContextManager)
-    @mock.patch('urllib.request.Request', git_api_state_for_test_github_sync.new_request())
-    def test_github_sync(self) -> None:
-        (
-            self.repo_sandbox
-                .new_branch('master')
-                .commit()
-                .push()
-                .new_branch('bar')
-                .commit()
-                .new_branch('bar2')
-                .commit()
-                .check_out("master")
-                .new_branch('foo')
-                .commit()
-                .push()
-                .new_branch('foo2')
-                .commit()
-                .check_out("master")
-                .new_branch('moo')
-                .commit()
-                .new_branch('moo2')
-                .commit()
-                .check_out("master")
-                .new_branch('snickers')
-                .push()
-        )
-        self.launch_command('discover', '-y')
-        (
-            self.repo_sandbox
-                .check_out("master")
-                .new_branch('mars')
-                .commit()
-                .check_out("master")
-        )
-        self.launch_command('github', 'sync')
-
-        expected_status_output = (
-            """
-            master *
-            |
-            o-bar (untracked)
-            |
-            o-foo
-            |
-            o-moo (untracked)
-            |
-            o-snickers  PR #7
-            """
-        )
-        self.assert_command(['status'], expected_status_output)
-
-        with pytest.raises(subprocess.CalledProcessError):
-            self.repo_sandbox.check_out("mars")
